markdown_helper.rb を作成し、ファイルを分割
