+++ app/helpers/application_helper.rb
@@ -1,47 +1,4 @@
 module ApplicationHelper
-  require 'redcarpet'
-  require 'coderay'
-  class HTMLwithCoderay < Redcarpet::Render::HTML
-    def block_code(code, language)
-      language = language.split(':')[0] if language.present?
-
-      lang = case language.to_s
-             when 'rb'
-               'ruby'
-             when 'yml'
-               'yaml'
-             when 'css'
-               'css'
-             when 'html'
-               'html'
-             when ''
-               # 空欄のままだと「Invalid id given:」エラー
-               'md'
-             else
-               language
-             end
-
-      CodeRay.scan(code, lang).div
-    end
-  end
-
-  def markdown(text)
-    html_render = HTMLwithCoderay.new(filter_html: true, hard_wrap: true)
-
-    options = {
-      autolink: true,
-      space_after_headers: true,
-      no_intra_emphasis: true,
-      fenced_code_blocks: true,
-      tables: true,
-      hard_wrap: true,
-      xhtml: true,
-      lax_html_blocks: true,
-      strikethrough: true
-    }
-    markdown = Redcarpet::Markdown.new(html_render, options)
-    markdown.render(text).html_safe
-  end
   
   def max_width
     devise_controller? ? 'mw-md' : 'mw-xl'
